CPM-1107: Rename Integration test
