+++ tests/back/Pim/Enrichment/Integration/Storage/Sql/Completeness/SaveProductCompletenessesIntegration.php
@@ -4,41 +4,17 @@ declare(strict_types=1);
 
 namespace AkeneoTest\Pim\Enrichment\Integration\Storage\Sql\Completeness;
 
-use Akeneo\Channel\Infrastructure\Component\Model\ChannelInterface;
 use Akeneo\Pim\Enrichment\Component\Product\Completeness\Model\ProductCompletenessWithMissingAttributeCodes;
 use Akeneo\Pim\Enrichment\Component\Product\Completeness\Model\ProductCompletenessWithMissingAttributeCodesCollection;
+use Akeneo\Test\Integration\Configuration;
 use Akeneo\Test\Integration\TestCase;
+use Doctrine\DBAL\Connection;
 use PHPUnit\Framework\Assert;
 use Ramsey\Uuid\UuidInterface;
 
-/**
- * @author    Mathias METAYER <mathias.metayer@akeneo.com>
- * @copyright 2019 Akeneo SAS (http://www.akeneo.com)
- * @license   http://opensource.org/licenses/osl-3.0.php  Open Software License (OSL 3.0)
- */
-class SaveProductCompletenessesIntegration extends TestCase
+final class SaveProductCompletenessesIntegration extends TestCase
 {
-    public function test_that_it_does_not_clear_existing_completenesses_and_missing_attributes_if_provided_completenesses_are_empty()
-    {
-        $productUuid = $this->createProduct('a_great_product');
-        Assert::assertCount(6, $this->getCompletenessesFromDB($productUuid));
-
-        $this->executeSave(new ProductCompletenessWithMissingAttributeCodesCollection($productUuid->toString(), []));
-        Assert::assertCount(6, $this->getCompletenessesFromDB($productUuid));
-    }
-
-    public function test_that_it_clears_non_existing_locales_channels()
-    {
-        $this->createChannel('my_channel', ['locales' => ['en_US', 'fr_FR']]);
-        $productUuid = $this->createProduct('a_great_product');
-        Assert::assertCount(8, $this->getCompletenessesFromDB($productUuid));
-
-        $this->removeChannel('my_channel');
-        $this->executeSave(new ProductCompletenessWithMissingAttributeCodesCollection($productUuid->toString(), []));
-        Assert::assertCount(6, $this->getCompletenessesFromDB($productUuid));
-    }
-
-    public function test_that_it_saves_completenesses_given_a_product_id()
+    public function test_that_it_saves_a_product_completenesses(): void
     {
         $productUuid = $this->createProduct('a_great_product');
         $collection = new ProductCompletenessWithMissingAttributeCodesCollection($productUuid->toString(), [
@@ -47,17 +23,21 @@ class SaveProductCompletenessesIntegration extends TestCase
         $this->executeSave($collection);
 
         $dbCompletenesses = $this->getCompletenessesFromDB($productUuid);
-        Assert::assertCount(6, $dbCompletenesses);
+        Assert::assertCount(1, $dbCompletenesses);
         Assert::assertEquals(
             [
-                'missing' => 0,
-                'required' => 5,
+                'ecommerce' => [
+                    'en_US' => [
+                        'missing' => 0,
+                        'required' => 5,
+                    ]
+                ]
             ],
-            $dbCompletenesses['ecommerce-en_US']
+            $dbCompletenesses,
         );
     }
 
-    public function test_that_it_saves_completenesses()
+    public function test_that_it_saves_completenesses(): void
     {
         $productUuid = $this->createProduct('a_great_product');
 
@@ -81,24 +61,27 @@ class SaveProductCompletenessesIntegration extends TestCase
         $this->executeSave($collection);
 
         $dbCompletenesses = $this->getCompletenessesFromDB($productUuid);
-        Assert::assertCount(6, $dbCompletenesses);
+        Assert::assertCount(2, $dbCompletenesses);
         Assert::assertEquals(
             [
-                'missing' => 1,
-                'required' => 5,
+                'ecommerce' => [
+                    'en_US' => [
+                        'missing' => 1,
+                        'required' => 5,
+                    ]
+                ],
+                'tablet' => [
+                    'fr_FR' => [
+                        'missing' => 6,
+                        'required' => 10,
+                    ]
+                ],
             ],
-            $dbCompletenesses['ecommerce-en_US']
-        );
-        Assert::assertEquals(
-            [
-                'missing' => 6,
-                'required' => 10,
-            ],
-            $dbCompletenesses['tablet-fr_FR']
+            $dbCompletenesses,
         );
     }
 
-    protected function getConfiguration()
+    protected function getConfiguration(): Configuration
     {
         return $this->catalog->useTechnicalCatalog();
     }
@@ -106,6 +89,11 @@ class SaveProductCompletenessesIntegration extends TestCase
     private function executeSave(ProductCompletenessWithMissingAttributeCodesCollection $completenesses): void
     {
         $this->get('akeneo.pim.enrichment.product.query.save_product_completenesses')->save($completenesses);
+    }
+
+    private function connection(): Connection
+    {
+        return $this->get('database_connection');
     }
 
     private function createProduct(string $identifier): UuidInterface
@@ -120,52 +108,11 @@ class SaveProductCompletenessesIntegration extends TestCase
     {
         $sql = <<<SQL
 SELECT completeness
-FROM pim_catalog_product_completeness
-WHERE pim_catalog_product_completeness.product_uuid = :productUuid
+FROM pim_catalog_product_completeness completeness
+WHERE product_uuid = :productUuid
 SQL;
-        $results = [];
-        $completenesses = \json_decode(
-            $this->get('database_connection')->executeQuery($sql, ['productUuid' => $productUuid->getBytes()])->fetchOne(),
-            true
-        );
+        $result = $this->connection()->executeQuery($sql, ['productUuid' => $productUuid->getBytes()])->fetchOne();
 
-        foreach ($completenesses as $channelCode => $completenessPerChannel) {
-            foreach ($completenessPerChannel as $localeCode => $completeness) {
-                $key = sprintf('%s-%s', $channelCode, $localeCode);
-                $results[$key] = $completeness;
-            }
-        }
-
-        return $results;
-    }
-
-    private function createChannel(string $code, array $data = []): ChannelInterface
-    {
-        $defaultData = [
-            'code' => $code,
-            'locales' => ['en_US'],
-            'currencies' => ['USD'],
-            'category_tree' => 'master',
-        ];
-        $data = array_merge($defaultData, $data);
-
-        $channel = $this->get('pim_catalog.repository.channel')->findOneByIdentifier($code);
-        if (null === $channel) {
-            $channel = $this->get('pim_catalog.factory.channel')->create();
-        }
-
-        $this->get('pim_catalog.updater.channel')->update($channel, $data);
-        $errors = $this->get('validator')->validate($channel);
-        Assert::assertCount(0, $errors);
-
-        $this->get('pim_catalog.saver.channel')->saveAll([$channel]);
-
-        return $channel;
-    }
-
-    private function removeChannel(string $code): void
-    {
-        $channel = $this->get('pim_catalog.repository.channel')->findOneByIdentifier($code);
-        $this->get('pim_catalog.remover.channel')->remove($channel);
+        return \json_decode($result, true);
     }
 }
